mk_profile_colemak: More useful mappings.
Physical Caps Lock is mapped to Right CTRL now, not Left CTRL. This is
because some stupid programs react to Right CTRL only.

Physical Left CTRL is mapped to Right Meta (Super R) so that we can
use KWin's tiling shortcuts. Hitting that key opens the application
menu as a bonus.